+++ app/src/main/java/com/techcourse/controller/LogoutController.java
@@ -1,6 +1,6 @@
 package com.techcourse.controller;
 
-import com.interface21.webmvc.servlet.stereotype.Controller;
+import com.interface21.context.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.view.ModelAndView;
