+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/ControllerScanner.java
@@ -1,6 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.webmvc.servlet.stereotype.Controller;
+import com.interface21.context.Controller;
 import java.lang.reflect.Constructor;
 import java.util.HashMap;
 import java.util.Map;
