+++ mvc/src/test/java/samples/TestController.java
@@ -1,6 +1,6 @@
 package samples;
 
-import com.interface21.webmvc.servlet.stereotype.Controller;
+import com.interface21.context.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
